requirements.txt
asgiref==3.5.0
black==22.1.0
click==8.0.3
Django==4.0.2
django-ckeditor==6.2.0
django-js-asset==2.0.0
django-multiselectfield==0.1.12
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.0
psycopg2==2.9.3
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.0
